Refactor engine tests into modular structure
