+++ debug.html
@@ -90,7 +90,7 @@
             runGuardianManagerTests,
             injectGuardianManagerFaults,
             runMeasureManagerIntegrationTest
-        } from './js/tests/engineTests.js';
+        } from './tests/index.js';
 
         document.addEventListener('DOMContentLoaded', () => {
             let gameEngine; // gameEngine 인스턴스를 스코프 밖으로 뺍니다.
