+++ tests/unit/mapManagerUnitTests.js
@@ -1,0 +1,3 @@
+export function runMapManagerUnitTests() {
+    console.warn('MapManager unit tests are not implemented yet.');
+}
