+++ tests/unit/measureManagerUnitTests.js
@@ -1,0 +1,3 @@
+export function runMeasureManagerUnitTests() {
+    console.warn('MeasureManager unit tests are not implemented yet.');
+}
