+++ tests/unit/uiManagerUnitTests.js
@@ -1,0 +1,3 @@
+export function runUIManagerUnitTests() {
+    console.warn('UIManager unit tests are not implemented yet.');
+}
